requirements.txt
# Lark Bot Dependencies

# Core dependencies
aiohttp>=3.8.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Crypto/Web3 dependencies
requests>=2.28.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx
